requirements.txt
# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1
python-multipart==0.0.9

# Database
motor==3.3.2
pymongo==4.6.1

# Cache
redis[hiredis]==5.0.1

# Task Queue
celery==5.3.6

# HTTP Client
httpx==0.27.0

# Web Scraping
beautifulsoup4==4.12.3
lxml==5.1.0
pdfplumber==0.10.4

# Authentication
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.1.0

# PDF Generation
reportlab==4.1.0

# AI / OpenAI
openai==1.12.0

# Notifications
twilio==8.13.0

# Google APIs
google-api-python-client==2.116.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# ML Models
numpy==1.26.4
pandas==2.2.0
scikit-learn==1.4.0

# Testing
pytest==8.0.1
pytest-asyncio==0.23.5
pytest-cov==4.1.0

# Dev Utilities
black==24.2.0
ruff==0.2.2
